Rename walkthrough internals to init
