+++ src/Installer/dotnetup/Commands/Init/InitCommand.cs
@@ -4,19 +4,19 @@
 using System.CommandLine;
 using Microsoft.DotNet.Tools.Bootstrapper.Commands.Shared;
 
-namespace Microsoft.DotNet.Tools.Bootstrapper.Commands.Walkthrough;
+namespace Microsoft.DotNet.Tools.Bootstrapper.Commands.Init;
 
 /// <summary>
 /// Runs the interactive initialization flow that installs the .NET SDK with defaults
 /// and records the user's path replacement preference to <c>dotnetup.config.json</c>.
 /// </summary>
-internal class WalkthroughCommand(ParseResult result) : InstallCommand(result)
+internal class InitCommand(ParseResult result) : InstallCommand(result)
 {
     protected override string GetCommandName() => "init";
 
     protected override int ExecuteCore()
     {
-        var workflows = new WalkthroughWorkflows(DotnetEnvironment, ChannelVersionResolver);
+        var workflows = new InitWorkflows(DotnetEnvironment, ChannelVersionResolver);
         workflows.FullIntroductionWalkthrough(this);
         return 0;
     }
