+++ src/Installer/dotnetup/Strings.resx
@@ -174,7 +174,7 @@
   <data name="TelemetryNotice" xml:space="preserve">
     <value>dotnetup collects usage data to help improve your experience. You can opt out by setting the DOTNET_CLI_TELEMETRY_OPTOUT environment variable to '1'. Learn more: https://aka.ms/dotnetup-telemetry</value>
   </data>
-  <data name="WalkthroughCommandDescription" xml:space="preserve">
+  <data name="InitCommandDescription" xml:space="preserve">
     <value>Configure or reconfigure dotnetup settings and install .NET through the interactive setup flow.</value>
   </data>
   <data name="PathPreferenceDotnetupDotnet" xml:space="preserve">
