+++ src/Installer/dotnetup/Telemetry/FirstRunNotice.cs
@@ -75,7 +75,7 @@ internal static class FirstRunNotice
     {
         // Write the telemetry notice to stderr using plain Console.Error to avoid
         // Spectre.Console rendering issues when stderr is not a terminal.
-        // The dotnetup banner is shown separately by WalkthroughCommand.
+        // The dotnetup banner is shown separately by InitCommand.
         // See: https://learn.microsoft.com/dotnet/core/compatibility/sdk/10.0/dotnet-cli-stderr-output
         Console.Error.WriteLine();
         Console.Error.WriteLine(Strings.TelemetryNotice);
