+++ test/dotnetup.Tests/DotnetBotBannerTests.cs
@@ -1,8 +1,8 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using Microsoft.DotNet.Tools.Bootstrapper.Commands.Init;
 using Spectre.Console;
-using Microsoft.DotNet.Tools.Bootstrapper.Commands.Walkthrough;
 
 namespace Microsoft.DotNet.Tools.Dotnetup.Tests;
